+++ src/pages/RecipientsList/RecipientsListPage.jsx
@@ -1,31 +1,11 @@
-import RecipientCard from './RecipientsCard';
-import useFetchData from '../../hooks/useFetchData';
-import { getRollingList } from '../../service/api';
+import RecipientsList from './RecipientsList';
 
-function RecipientsList() {
-  const { data: rollingListData } = useFetchData(getRollingList, []);
-  const rollingList = rollingListData.results;
-
+function RecipientsListPage() {
   return (
-    <ul>
-      {rollingList.map((item) => (
-        <li key={item.id}>
-          <RecipientCard
-            id={item.id}
-            name={item.name}
-            bgColor={item.backgroundColor}
-            bgImage={item.backgroundImageURL}
-            totalMessage={{
-              recentMessages: item.recentMessages,
-              messageCount: item.messageCount,
-              direction: 'column',
-            }}
-            emojiList={item.topReactions}
-          />
-        </li>
-      ))}
-    </ul>
+    <>
+      <RecipientsList />
+    </>
   );
 }
 
-export default RecipientsList;
+export default RecipientsListPage;
